+++ python/ray/serve/types.py
@@ -1,0 +1,8 @@
+from uuid import UUID
+
+BackendTag = str
+EndpointTag = str
+ReplicaTag = str
+NodeId = str
+GoalId = UUID
+Duration = float
